Correção de casas decimais

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -15,10 +15,10 @@ public class Calculadora {
 		System.out.println("Digite o segundo número: ");
 		b = scan.nextInt();
 		
-		int soma = soma(a, b);
-		int subtracao = subtracao(a, b);
-		int multiplicacao = multiplicacao(a, b);
-		int divisao = divisao(a, b);
+		double soma = soma(a, b);
+		double subtracao = subtracao(a, b);
+		double multiplicacao = multiplicacao(a, b);
+		double divisao = divisao(a, b);
 		
 		System.out.println("Soma " + soma);
 		System.out.println("Subtracao " + subtracao);
@@ -27,16 +27,16 @@ public class Calculadora {
 
 	}
 	
-	public static int soma(int a, int b) {
+	public static double soma(double a, double b) {
 		return a + b;
 	}
-	public static int subtracao(int a, int b) {
+	public static double subtracao(double a, double b) {
 		return a - b;
 	}
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a / b;
 	}
-	public static int multiplicacao(int a, int b) {
+	public static double multiplicacao(double a, double b) {
 		return a * b;
 	}
